Created common.py with common functions
Used by main and by utilities, placed here to avoid circular referencing
(so main could call utilities w/o utilities importing from main)
Mainly has logging-related functions, relocated from main.
Removed the respective functions from main.
