+++ arc/__init__.py
@@ -4,6 +4,7 @@
 import arc_exceptions
 import main
 from main import ARC
+import common
 import parser
 import plotter
 import processor
